Clean up Scheduler forks
