--- packages/scheduler/npm/unstable_no_dom.js
@@ -1,7 +1,0 @@
-'use strict';
-
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/scheduler-unstable_no_dom.production.min.js');
-} else {
-  module.exports = require('./cjs/scheduler-unstable_no_dom.development.js');
-}
