+++ packages/scheduler/package.json
@@ -28,7 +28,6 @@
     "tracing.js",
     "tracing-profiling.js",
     "unstable_mock.js",
-    "unstable_no_dom.js",
     "unstable_post_task.js",
     "cjs/",
     "umd/"
