+++ packages/scheduler/src/__tests__/Scheduler-test.js
@@ -32,9 +32,7 @@ let NormalPriority;
 describe('SchedulerBrowser', () => {
   beforeEach(() => {
     jest.resetModules();
-
-    // Un-mock scheduler
-    jest.mock('scheduler', () => require.requireActual('scheduler'));
+    jest.unmock('scheduler');
 
     runtime = installMockBrowserRuntime();
     performance = global.performance;
@@ -70,11 +68,6 @@ describe('SchedulerBrowser', () => {
 
     const window = {};
     global.window = window;
-
-    // TODO: Scheduler no longer requires these methods to be polyfilled. But
-    // maybe we want to continue warning if they don't exist, to preserve the
-    // option to rely on it in the future?
-    window.requestAnimationFrame = window.cancelAnimationFrame = () => {};
 
     window.setTimeout = (cb, delay) => {
       const id = timerIDCounter++;
