+++ packages/scheduler/src/__tests__/SchedulerPostTask-test.js
@@ -30,8 +30,6 @@ let IdlePriority;
 describe('SchedulerPostTask', () => {
   beforeEach(() => {
     jest.resetModules();
-
-    // Un-mock scheduler
     jest.mock('scheduler', () =>
       require.requireActual('scheduler/unstable_post_task'),
     );
