--- packages/scheduler/unstable_no_dom.js
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-'use strict';
-
-export * from './src/forks/SchedulerNoDOM';
