+++ scripts/rollup/bundles.js
@@ -724,39 +724,6 @@ const bundles = [
     externals: [],
   },
 
-  /******* React Scheduler Post Task Only (experimental) *******/
-  {
-    bundleTypes: [
-      NODE_DEV,
-      NODE_PROD,
-      FB_WWW_DEV,
-      FB_WWW_PROD,
-      FB_WWW_PROFILING,
-    ],
-    moduleType: ISOMORPHIC,
-    entry: 'scheduler/unstable_post_task_only',
-    global: 'SchedulerPostTaskOnly',
-    externals: [],
-  },
-
-  /******* React Scheduler No DOM (experimental) *******/
-  {
-    bundleTypes: [
-      NODE_DEV,
-      NODE_PROD,
-      FB_WWW_DEV,
-      FB_WWW_PROD,
-      FB_WWW_PROFILING,
-      RN_FB_DEV,
-      RN_FB_PROD,
-      RN_FB_PROFILING,
-    ],
-    moduleType: ISOMORPHIC,
-    entry: 'scheduler/unstable_no_dom',
-    global: 'SchedulerNoDOM',
-    externals: [],
-  },
-
   /******* Jest React (experimental) *******/
   {
     bundleTypes: [NODE_DEV, NODE_PROD],
